+++ NIDAQ_OpenScanSettings.h
