+++ meson.build
@@ -62,8 +62,8 @@ src = files(
     'Detector.c',
     'NIDAQ_OpenScanDevice.c',
     'NIDAQ_OpenScanModule.c',
+    'NIDAQ_OpenScanSettings.c',
     'OScNIDAQ.c',
-    'OScNIDAQSettings.c',
     'ParkUnpark.c',
     'Scanner.c',
     'Waveform.c',
